#2003: extract fd pinning + degraded-path reader into bpf_map/pin.rs
Final code-motion step of the bpf_map/mod.rs split. Lifts the only items
that talk to bpffs pin paths directly into a new bpf_map/pin.rs, leaving
mod.rs as the coordinator over the session-map publish/delete core and
the conntrack-mirror orchestrators.

Moved verbatim (no signature or logic change): the OwnedFd RAII wrapper
(bpf_obj_get on open, libc::close on Drop), the pinned-map path constant
DEGRADED_PATH_STATS_PIN_PATH, the DEGRADED_PATH_REASON_NAMES table, and
the debug-gated read_degraded_path_stats reader. The cfg gating
(feature = "debug-log" / any(test, feature = "debug-log")) moves with the
items unchanged.

OwnedFd keeps its original pub(super) visibility (= pub(in crate::afxdp))
via an explicit declaration and is re-exported with `pub(in crate::afxdp)
use pin::*;`, so the external `crate::afxdp::bpf_map::OwnedFd` paths in
coordinator/bpf_maps.rs and coordinator/worker_manager.rs resolve to the
exact same item at the exact same visibility. The pre-existing
private_interfaces lint on the pub(crate) BpfMaps fields that hold
OwnedFd is unchanged in severity — only the type's display path moves
from bpf_map::OwnedFd to bpf_map::pin::OwnedFd.

mod.rs drops from ~1063 to ~640 LOC. Public API surface of
crate::afxdp::bpf_map is bit-identical (same names, same
pub(in crate::afxdp) visibility).

Validation: cargo build --release clean; cargo test --release passes
(2057 lib + 46 + 8 + 16 + 1 across the other test binaries, 0 failed;
the relocated bpf_map_tests.rs resolves every moved item by bare name).
